+++ packages/scm/src/browser/scm-tree-widget.tsx
@@ -561,7 +561,7 @@ export namespace ScmElement {
 
 export class ScmResourceComponent extends ScmElement<ScmResourceComponent.Props> {
 
-    override render(): JSX.Element | undefined {
+    override render(): React.JSX.Element | undefined {
         const { hover } = this.state;
         const { model, treeNode, colors, parentPath, sourceUri, decoration, labelProvider, menus, contextKeys, caption, isLightTheme } = this.props;
         const resourceUri = new URI(sourceUri);
@@ -682,7 +682,7 @@ export namespace ScmResourceComponent {
 
 export class ScmResourceGroupElement extends ScmElement<ScmResourceGroupComponent.Props> {
 
-    override render(): JSX.Element {
+    override render(): React.JSX.Element {
         const { hover } = this.state;
         const { model, treeNode, menus, contextKeys, caption } = this.props;
         return <div className={`theia-header scm-theia-header ${TREE_NODE_SEGMENT_GROW_CLASS}`}
@@ -732,7 +732,7 @@ export namespace ScmResourceGroupComponent {
 
 export class ScmResourceFolderElement extends ScmElement<ScmResourceFolderElement.Props> {
 
-    override render(): JSX.Element {
+    override render(): React.JSX.Element {
         const { hover } = this.state;
         const { model, treeNode, sourceUri, labelProvider, menus, contextKeys, caption } = this.props;
         const sourceFileStat = FileStat.dir(sourceUri);